Rename hydrate function and what it handles
